Move special browser tests into its own test file
