+++ test/index.js
@@ -199,113 +199,4 @@ describe('csp middleware', function () {
     assert.equal(csp.name, 'csp')
     assert.equal(csp({ directives: POLICY }).name, 'csp')
   })
-
-  describe('special browsers', function () {
-    it('sets the header properly for Firefox 22', function (done) {
-      var app = use({
-        directives: {
-          defaultSrc: ["'self'"],
-          connectSrc: ['connect.com']
-        }
-      })
-
-      request(app).get('/').set('User-Agent', AGENTS['Firefox 22'].string)
-      .end(function (err, res) {
-        if (err) { return done(err) }
-
-        assert.deepEqual(parseCsp(res.headers['x-content-security-policy']), {
-          'default-src': ["'self'"],
-          'xhr-src': ['connect.com']
-        })
-
-        done()
-      })
-    })
-
-    ;[
-      'Safari 4.1',
-      'Safari 5.1 on OS X',
-      'Safari 5.1 on Windows Server 2008'
-    ].forEach(function (browser) {
-      it("doesn't set the property for " + browser, function (done) {
-        var app = use({ directives: POLICY })
-
-        request(app).get('/').set('User-Agent', AGENTS[browser].string)
-        .end(function (err, res) {
-          if (err) { return done(err) }
-
-          assert.equal(res.header['x-webkit-csp'], undefined)
-          assert.equal(res.header['content-security-policy'], undefined)
-          assert.equal(res.header['x-content-security-policy'], undefined)
-
-          done()
-        })
-      })
-    })
-
-    it('lets you disable Android', function (done) {
-      var app = use({
-        disableAndroid: true,
-        directives: {
-          defaultSrc: ['a.com']
-        }
-      })
-
-      request(app).get('/').set('User-Agent', AGENTS['Android 4.4.3'].string)
-      .end(function (err, res) {
-        if (err) { return done(err) }
-
-        assert.equal(res.header['x-webkit-csp'], undefined)
-        assert.equal(res.header['content-security-policy'], undefined)
-        assert.equal(res.header['x-content-security-policy'], undefined)
-
-        done()
-      })
-    })
-
-    it("appends connect-src 'self' in iOS Chrome when connect-src is already defined", function (done) {
-      var app = use({
-        directives: {
-          connectSrc: ['connect.com']
-        }
-      })
-      var iosChrome = AGENTS['iOS Chrome 40']
-
-      request(app).get('/').set('User-Agent', iosChrome.string)
-      .end(function (err, res) {
-        if (err) { return done(err) }
-
-        var header = iosChrome.header.toLowerCase()
-        var connectSrc = parseCsp(res.headers[header])['connect-src'].sort()
-        assert.deepEqual(connectSrc, ["'self'", 'connect.com'])
-
-        done()
-      })
-    })
-
-    it("adds connect-src 'self' in iOS Chrome when connect-src is undefined", function (done) {
-      var app = use({
-        directives: {
-          styleSrc: ["'self'"]
-        }
-      })
-      var iosChrome = AGENTS['iOS Chrome 40']
-
-      request(app).get('/').set('User-Agent', iosChrome.string)
-      .expect(iosChrome.header, /connect-src 'self'/)
-      .end(done)
-    })
-
-    it("does nothing in iOS Chrome if connect-src 'self' is defined", function (done) {
-      var app = use({
-        directives: {
-          connectSrc: ['somedomain.com', "'self'"]
-        }
-      })
-      var iosChrome = AGENTS['iOS Chrome 40']
-      request(app).get('/').set('User-Agent', iosChrome.string)
-      .expect(iosChrome.header, "connect-src somedomain.com 'self'")
-      .end(done)
-    })
-  })
 })
